+++ src/client/datascience/activation.ts
@@ -12,7 +12,7 @@ import { IDisposableRegistry } from '../common/types';
 import { debounceAsync, swallowExceptions } from '../common/utils/decorators';
 import { sendTelemetryEvent } from '../telemetry';
 import { JupyterDaemonModule, Telemetry } from './constants';
-import { ActiveEditorContextService } from './context/activeEditorContext';
+import { ActiveEditorContextService } from './telemetry/activeEditorContext';
 import { JupyterInterpreterService } from './jupyter/interpreter/jupyterInterpreterService';
 import { KernelDaemonPreWarmer } from './kernel-launcher/kernelDaemonPreWarmer';
 import { INotebookCreationTracker, INotebookEditorProvider } from './types';
