+++ src/client/datascience/jupyter/kernels/kernelSelector.ts
@@ -20,8 +20,8 @@ import { PythonEnvironment } from '../../../pythonEnvironments/info';
 import { captureTelemetry, IEventNamePropertyMapping, sendTelemetryEvent } from '../../../telemetry';
 import { sendNotebookOrKernelLanguageTelemetry } from '../../common';
 import { Commands, Telemetry } from '../../constants';
-import { sendKernelListTelemetry } from '../../context/kernelTelemetry';
-import { sendKernelTelemetryEvent, trackKernelResourceInformation } from '../../context/telemetry';
+import { sendKernelListTelemetry } from '../../telemetry/kernelTelemetry';
+import { sendKernelTelemetryEvent, trackKernelResourceInformation } from '../../telemetry/telemetry';
 import { IKernelFinder, IpyKernelNotInstalledError } from '../../kernel-launcher/types';
 import { isPythonNotebook } from '../../notebook/helpers/helpers';
 import { getInterpreterInfoStoredInMetadata } from '../../notebookStorage/baseModel';
