+++ src/client/datascience/notebook/kernelProvider.ts
@@ -17,8 +17,8 @@ import { StopWatch } from '../../common/utils/stopWatch';
 import { captureTelemetry } from '../../telemetry';
 import { sendNotebookOrKernelLanguageTelemetry } from '../common';
 import { Telemetry } from '../constants';
-import { sendKernelListTelemetry } from '../context/kernelTelemetry';
-import { getErrorClassification, sendKernelTelemetryEvent, trackKernelResourceInformation } from '../context/telemetry';
+import { sendKernelListTelemetry } from '../telemetry/kernelTelemetry';
+import { getErrorClassification, sendKernelTelemetryEvent, trackKernelResourceInformation } from '../telemetry/telemetry';
 import { areKernelConnectionsEqual, isLocalLaunch } from '../jupyter/kernels/helpers';
 import { KernelSelectionProvider } from '../jupyter/kernels/kernelSelections';
 import { KernelSelector } from '../jupyter/kernels/kernelSelector';
