+++ src/client/datascience/telemetry/errorClassificationRegistration.ts
