+++ src/client/datascience/telemetry/serviceRegistry.ts
